rename old anames to new one
