+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/api/ext/provider/kononnullabletype/snippet/declaration-has-simple-type.kttxt
