+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/api/ext/provider/konullabletype/KoNullableTypeProviderExtTest.kt
@@ -1,26 +1,39 @@
-package com.lemonappdev.konsist.api.ext.provider.kotype
+package com.lemonappdev.konsist.api.ext.provider.konullabletype
 
 import com.lemonappdev.konsist.TestSnippetProvider
 import com.lemonappdev.konsist.api.ext.koscope.declarationsOf
-import com.lemonappdev.konsist.api.provider.KoNonNullableTypeProvider
+import com.lemonappdev.konsist.api.provider.KoNullableTypeProvider
 import com.lemonappdev.konsist.testdata.SampleClass
-import com.lemonappdev.konsist.testdata.SampleType
 import hasTypeOf
 import org.amshove.kluent.assertSoftly
 import org.amshove.kluent.shouldBeEqualTo
 import org.junit.jupiter.api.Test
 
-class KoNonNullableTypeProviderExtTest {
+class KoNullableTypeProviderExtTest {
     @Test
-    fun `declaration-has-simple-type`() {
+    fun `declaration-has-no-type`() {
         // given
-        val sut = getSnippetFile("declaration-has-simple-type")
-            .declarationsOf<KoNonNullableTypeProvider>()
+        val sut = getSnippetFile("declaration-has-no-type")
+            .declarationsOf<KoNullableTypeProvider>()
             .first()
 
         // then
         assertSoftly(sut) {
-            hasTypeOf<String>() shouldBeEqualTo true
+            hasTypeOf<Int>() shouldBeEqualTo false
+            hasTypeOf<SampleClass>() shouldBeEqualTo false
+        }
+    }
+
+    @Test
+    fun `declaration-has-simple-type`() {
+        // given
+        val sut = getSnippetFile("declaration-has-simple-type")
+            .declarationsOf<KoNullableTypeProvider>()
+            .first()
+
+        // then
+        assertSoftly(sut) {
+            hasTypeOf<Int>() shouldBeEqualTo true
             hasTypeOf<SampleClass>() shouldBeEqualTo false
         }
     }
@@ -29,16 +42,16 @@ class KoNonNullableTypeProviderExtTest {
     fun `declaration-has-complex-type`() {
         // given
         val sut = getSnippetFile("declaration-has-complex-type")
-            .declarationsOf<KoNonNullableTypeProvider>()
+            .declarationsOf<KoNullableTypeProvider>()
             .first()
 
         // then
         assertSoftly(sut) {
-            hasTypeOf<SampleType>() shouldBeEqualTo true
+            hasTypeOf<SampleClass>() shouldBeEqualTo true
             hasTypeOf<Int>() shouldBeEqualTo false
         }
     }
 
     private fun getSnippetFile(fileName: String) =
-        TestSnippetProvider.getSnippetKoScope("api/ext/provider/kotype/snippet/", fileName)
+        TestSnippetProvider.getSnippetKoScope("api/ext/provider/konullabletype/snippet/", fileName)
 }
