+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/api/ext/provider/konullabletype/snippet/declaration-has-no-type.kttxt
