+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/api/ext/provider/konullabletype/snippet/declaration-has-simple-type.kttxt
